refactor project to break .h into .h and .c
breakout for declaration and definition in separate files
