+++ inc/amp.h
@@ -1,47 +1,8 @@
 # ifndef __CLIFE_AMP_H__
 # define __CLIFE_AMP_H__
+
 #include<stdio.h>
-void printASCIImap(char *location) {
-  printf("You are in %s.\n\n", location);
-  printf("+------CLIfe-----Aritrean-Circle-------------------------+\n");
-  printf("|      \\                                        |        |\n");
-  printf("|       |                                       |        |\n");
-  printf("|        \\_/\\         /^\\__                     |        |\n");
-  printf("|        /   \\_      /     \\__                  |        |\n");
-  printf("|      _/      \\_ _ /\\        \\__               |        |\n");
-  printf("|     /          v   |           \\__            |        |\n");
-  printf("|    /                \\ x<-Nova A   \\_          |-----^--|\n");
-  printf("|   /                  | ..x<-Majkius \\         |        |\n");
-  printf("| _/                    \\  .           \\        |        |\n");
-  printf("|/            Orar->x... |  .           \\__     |        |\n");
-  printf("|                      .  \\  .             \\ ,--'        |\n");
-  printf("|   LIBERIUM           ..../  .x<-Metron    \\|           |\n");
-  printf("|   -;-;-;-;             _|                  |           |\n");
-  printf("|                        |                   |           |\n");
-  printf("|                       /                    |           |\n");
-  printf("|                      |                     |           |\n");
-  printf("|                     _|                   __|           |\n");
-  printf("|                    /                     |             |\n");
-  printf("|                   /     ARITREA         /              |\n");
-  printf("|                  /      -;-;-;-       _|               |\n");
-  printf("|               __|                     /                |\n");
-  printf("|_    ____    _|   \\                   /                 |\n");
-  printf("| \\__/    \\__/      \\_                 \\_                |\n");
-  printf("|                     \\_                 |_              |\n");
-  printf("|                       |                  |             |\n");
-  printf("|                       |                  \\             |\n");
-  printf("|                        \\                 _)            |\n");
-  printf("|                       __\\              _/ \\            |\n");
-  printf("|                      /                |    \\_          |\n");
-  printf("|                      \\              _/      |          |\n");
-  printf("|                       \\_         __/       /           |\n");
-  printf("|                      /  \\      _/         |__          |\n");
-  printf("|                     )    |_   _/             \\_        |\n");
-  printf("|                    )       vvv                 \\       |\n");
-  printf("|         /^_/\\_    /      _/                     \\_     |\n");
-  printf("|        /     \\   /      /                         )    |\n");
-  printf("|       /       |_/      (                           \\_  |\n");
-  printf("|      |                 |                            |  |\n");
-  printf("+--------------------------------------------------------+\n");
-}
+
+void printASCIImap(char *location);
+
 # endif
