+++ inc/defs.h
@@ -1,0 +1,11 @@
+# ifndef __CLIFE_DEFS_H__
+# define __CLIFE_DEFS_H__
+
+#define NAME "arch/svf/nom.rtf" // Name
+#define ROLE "arch/svf/rol.rtf" // Role
+#define GOLD "arch/svf/gol.rtf" // Money
+#define POSI "arch/svf/loc.rtf" // Player's position
+#define MOTI "arch/svf/mot.rtf" // Motivation
+#define MANU "arch/man/" // Help files
+
+# endif
