+++ inc/money.h
@@ -1,12 +1,10 @@
 # ifndef __CLIFE_MONEY_H__
 # define __CLIFE_MONEY_H__
-#include<stdio.h>
 
-int addMoney(int i, int j) { // i = old money, j = money to be added
-  return i + j;
-}
+#include <stdio.h>
 
-int removeMoney(int i, int j) { // // i = old money, j = money to be removed
-  return i - j;
-}
+int addMoney(int i, int j); // i = old money, j = money to be added
+
+int removeMoney(int i, int j); // // i = old money, j = money to be removed
+
 # endif
