+++ inc/structs.h
@@ -1,5 +1,10 @@
 # ifndef __CLIFE_STRUCTS_H__
 # define __CLIFE_STRUCTS_H__
+
+# include <string.h>
+
+# include "env.h"
+
 struct DATA { // for the data read from the save file
   int gold;
   int moti;
@@ -19,32 +24,10 @@ struct city {
   char *buildings[12];
   char key[12];
 } Majkius, NovaA, Metron, Orar;
-void setAll() { // Get rid of reading dozens of files and replace it with a nice struct
-  Majkius.name = "Majkius";
-  Majkius.buildings[0] = "Church";
-  Majkius.key[0] = 'r';
 
-  NovaA.name = "NovaA";
-  NovaA.buildings[0] = "Plateau";
-  NovaA.key[0] = 'a';
+void setAll(); // Get rid of reading dozens of files and replace it with a nice struct
 
-  Metron.name = "Metron";
-
-  Orar.name = "Orar";
-}
-
-void detStruct(char *location, char *country, int la) { // la = lookAround? <- determines if the command is lookAround or inspectDialog
-  if(!strcmp(country,"Aritrea")) {
-    if(!strcmp(location,"Majkius")) if(la==1) lookAround(location, country, Majkius.buildings); else inspectDialog(location, country, Majkius.buildings, Majkius.key);
-    if(!strcmp(location,"Nova A")) if(la==1) lookAround(location, country, NovaA.buildings); else inspectDialog(location, country, NovaA.buildings, NovaA.key);
-    if(!strcmp(location,"Metron")) if(la==1) lookAround(location, country, Metron.buildings); else inspectDialog(location, country, Metron.buildings, Metron.key);
-  }
-  if(!strcmp(country,"Liberium")) {
-    if(!strcmp(location,"Orar")) if(la==1) lookAround(location, country, Orar.buildings); else inspectDialog(location, country, Orar.buildings, Orar.key);
-  }
-  return;
-}
-
+void detStruct(char *location, char *country, int la); // la = lookAround? <- determines if the command is lookAround or inspectDialog
 
 // Keymap
 /* a - Center of a city
