+++ src/money.c
@@ -1,0 +1,9 @@
+#include "money.h"
+
+int addMoney(int i, int j) { // i = old money, j = money to be added
+  return i + j;
+}
+
+int removeMoney(int i, int j) { // // i = old money, j = money to be removed
+  return i - j;
+}
